+++ app/components/speakers/SpeakerPageBlock.vue
